Add iCDG's "SPSId" metadata
These are unique IDs that index further metadata in a database managed by
iCDG. they are not expected to interfere with 'repo' script behavior
This change is required to allow iCDG to rebase their sofia mainline
manifests onto 1A manifests. The change was agreed by Rusty and others in
Munich this week.

Change-Id: I2e95839cc3abd3ac344d5b0706951a3ab5e6d68b
Signed-off-by: Matt Gumbel <matthew.k.gumbel@intel.com>
Signed-off-by: Stefan Stanacar <stefan.stanacar@intel.com>
Reviewed-on: https://android.intel.com:443/376509